# swagger.yaml
swagger: '2.0'
info:
  title: Swagchat RESTful API
  description: ""
  version: 0.1.0
  contact:
    email: shinichi.minobe@gmail.com
  license:
    name: MIT
    url: https://github.com/fairway-corp/swagchat-api/blob/master/LICENSE
host: localhost:9000
basePath: /v0
schemes:
- http
paths:
  /users:
    post:
      summary: Create user item.
      produces:
      - application/json
      consumes:
      - application/json
      parameters:
      - in: body
        name: user item
        description: Create user item json
        required: true
        schema:
          $ref: '#/definitions/postUser'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/user'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/problemDetailForBadGateway'
          examples:
            application/json: |-
              {
                "title": "Request parameter error. (Create user item)",
                "status": 400,
                "errorName": "invalid-param",
                "invalidParams": [
                  {
                    "name": "name",
                    "reason": "name is required, but it's empty."
                  }
                ]
              }
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemDetailForInternalServerError'
          examples:
            application/json: |-
              {
                "title": "An error occurred while creating user item.",
                "status": 500,
                "detail": "UNIQUE constraint failed: user.user_id",
                "errorName": "database-error"
              }
    get:
      summary: Get user list.
      produces:
      - application/json
      responses:
        200:
          description: User list
          schema:
            type: object
            items:
              $ref: '#/definitions/users'
          examples:
            application/json: |-
              {
                "users": [
                  {
                    "id": 1,
                    "userId": "custom-user-id-0001",
                    "name": "custom user 0001",
                    "unreadCount": 0,
                    "customData": {"key": "value"},
                    "created": 1488294000000000000,
                    "modified": 1488294000000000000 
                  }
                ]
              }
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemDetailForInternalServerError'
          examples:
            application/json: |-
              {
                "title": "An error occurred while creating user item.",
                "status": 500,
                "detail": "UNIQUE constraint failed: user.user_id",
                "errorName": "database-error"
              }
...
  /users/{userId}/rooms:
    get:
      summary: Get user's room list
      produces:
      - application/json
      parameters:
      - in: path
        name: userId
        description: User ID
        required: true
        type: string
        x-example: custom-user-id-0001
      responses:
        200:
          description: User's room list
          schema:
            type: array
            items:
              $ref: '#/definitions/room'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/problemDetailForBadGateway'
          examples:
            application/json: |-
              {
                "title": "Request parameter error. (Create user item)",
                "status": 400,
                "errorName": "invalid-param",
                "invalidParams": [
                  {
                    "name": "name",
                    "reason": "name is required, but it's empty."
                  }
                ]
              }
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemDetailForInternalServerError'
          examples:
            application/json: |-
              {
                "title": "An error occurred while creating user item.",
                "status": 500,
                "detail": "UNIQUE constraint failed: user.user_id",
                "errorName": "database-error"
              }
  /rooms:
    post:
      summary: Create room item.
      produces:
      - application/json
      consumes:
      - application/json
      parameters:
      - in: body
        name: room item
        description: Create room item json
        required: true
        schema:
          $ref: '#/definitions/postRoom'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/room'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/problemDetailForBadGateway'
          examples:
            application/json: |-
              {
                "title": "Request parameter error. (Create user item)",
                "status": 400,
                "errorName": "invalid-param",
                "invalidParams": [
                  {
                    "name": "name",
                    "reason": "name is required, but it's empty."
                  }
                ]
              }
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemDetailForInternalServerError'
          examples:
            application/json: |-
              {
                "title": "An error occurred while creating user item.",
                "status": 500,
                "detail": "UNIQUE constraint failed: user.user_id",
                "errorName": "database-error"
              }
    get:
      summary: Get room list.
      produces:
      - application/json
      responses:
        200:
          description: Room list
          schema:
            type: array
            items:
              $ref: '#/definitions/room'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemDetailForInternalServerError'
          examples:
            application/json: |-
              {
                "title": "An error occurred while creating user item.",
                "status": 500,
                "detail": "UNIQUE constraint failed: user.user_id",
                "errorName": "database-error"
              }
...
definitions:
  user:
    type: object
    required:
    - id
    - userId
    - name
    - unreadCount
    - customData
    - created
    - modified
    properties:
      id:
        type: integer
        example: 1
      userId:
        type: string
        example: custom-user-id-0001
      name:
        type: string
        example: custom user 0001
      pictureUrl:
        type: string
        example: http://example.com/img/rick.png
      informationUrl:
        type: string
        example: http://example.com/user/rick
      unreadCount:
        type: integer
        example: 0
      customData:
        $ref: '#/definitions/sampleCustomData'
      created:
        type: integer
        example: 1488294000000000000
      modified:
        type: integer
        example:
          key: value
  users:
    type: object
    required:
    - users
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/user'
  postUser:
    type: object
    required:
    - name
    properties:
      userId:
        type: string
        pattern: "[0-9a-zA-Z-]"
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      name:
        type: string
        example: rick
      pictureUrl:
        type: string
        example: http://example.com/img/rick.png
      informationUrl:
        type: string
        example: http://example.com/user/rick
      customData:
        $ref: '#/definitions/sampleCustomData'
  sampleCustomData:
    type: object
    properties:
      key:
        type: string
        example: value
  room:
    type: object
    required:
    - id
    - roomId
    - name
    - customData
    - isPublic
    - lastMessage
    - lastMessageUpdated
    - created
    - modified
    properties:
      id:
        type: integer
        example: 1
      roomId:
        type: string
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      name:
        type: string
        example: rick
      pictureUrl:
        type: string
        example: http://example.com/img/rick.png
      informationUrl:
        type: string
        example: http://example.com/user/rick
      customData:
        $ref: '#/definitions/sampleCustomData'
      isPublic:
        type: boolean
        example: true
      lastMessage:
        type: string
        example: Hello, World
      lastMessageUpdated:
        type: integer
        example: 1488294000000000000
      created:
        type: integer
        example: 1488294000000000000
      modified:
        type: integer
        example: 1488294000000000000
  postRoom:
    type: object
    required:
    - name
    properties:
      roomId:
        type: string
        pattern: "[0-9a-zA-Z-]"
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      name:
        type: string
        example: rick's room
      pictureUrl:
        type: string
        example: http://example.com/img/rick-room.png
      informationUrl:
        type: string
        example: http://example.com/room/rick
      customData:
        $ref: '#/definitions/sampleCustomData'
      isPublic:
        type: boolean
        example: true
...
  problemDetailForBadGateway :
    type: object
    required:
    - title
    - status
    - errorName
    properties:
      title:
        type: string
        example: Request parameter error. (Create room item).
      status:
        type: integer
        example: 400
      errorName:
        type: string
        example: invalid-json
      invalidParams: 
        type: array
        items:
          $ref: '#/definitions/invalidParam'
  problemDetailForInternalServerError :
    type: object
    required:
    - title
    - status
    - errorName
    - detail
    properties:
      title:
        type: string
        example: Your request parameters didn't validate.
      status:
        type: integer
        example: 500
      errorName:
        type: string
        example: database-error
      detail:
        type: string
        example: no rows in result set.
  invalidParam:
    type: object
    properties:
      name:
        type: string
        example: name
      reason:
        type: string
        example: name is required, but it's empty.